Fix classInView plugin not accepting floats

+++ src/plugins/classInView/classInViewPlugin.ts
@@ -12,7 +12,9 @@ const isEdge = getUserAgent() === "edge";
 const targets = document.querySelectorAll<HTMLElement>(`[${classInViewAttr}]`);
 
 for (const target of targets) {
-	const threshold = parseInt(target.dataset[classInViewThresholdDataKey] || "");
+	const threshold = parseFloat(
+		target.dataset[classInViewThresholdDataKey] || ""
+	);
 	if (!isInViewport(target)) toggleTargetOff(target);
 
 	inView(
